# Improve performance of type-class dict lookup